requirements.txt
streamlit==1.36.0
streamlit-option-menu==0.3.13
pandas==2.0.3
numpy==1.24.3
requests==2.32.5
python-dotenv==1.0.1
groq==0.11.0
langdetect==1.0.9
mysql-connector-python==8.2.0
setuptools==69.0.0
